pyproject.toml
[project]
name = "ai-service-mesh"
version = "0.1.0"
description = "AI service mesh demo services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "httpx>=0.27",
    "numpy==1.26.4",
    "pydantic>=2.6",
    "pillow>=10.2",
    "sentence-transformers>=2.6",
    "torch==2.2.2",
    "torchvision==0.17.2",
    "transformers>=4.39",
    "uvicorn>=0.27",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
